refactor : move files to component
