+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/design_system/component/BottomSheet.kt
@@ -1,4 +1,4 @@
-package com.cairosquad.evolvefit.design_system.composables
+package com.cairosquad.evolvefit.design_system.component
 
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.padding
